+++ modules/platforms/__init__.py
@@ -1,7 +1,7 @@
 """
 Init script to handle window manager.
 """
-from .factory import get_window_manager as _get_window_manager
+from modules.platforms.factory import get_window_manager as _get_window_manager
 
 
 win = _get_window_manager()