+++ modules/platforms/platforms.py
@@ -6,8 +6,9 @@ Functions:
 
 """
 
-import sys
 import logging
+import sys
+
 
 log = logging.getLogger(__name__)